feat: add main menu as base for new game modes
